+++ packages/core/src/service/openapi.ts
@@ -5,12 +5,13 @@ import {
   type RemoteConfig,
   getValidData,
   setCachedData,
+  getLocalData,
+  setLocalData,
 } from '@onflow/frw-data-model';
 import type { Account as FclAccount } from '@onflow/typedefs';
 import BigNumber from 'bignumber.js';
 import dayjs from 'dayjs';
 
-import storage from '@onflow/frw-extension-shared/storage';
 import {
   DEFAULT_CURRENCY,
   Period,
@@ -56,12 +57,13 @@ import {
   consoleLog,
   getPeriodFrequency,
   getPriceProvider,
+  consoleInfo,
 } from '@onflow/frw-shared/utils';
 
 import { findKeyAndInfo } from '../utils';
 import {
   googleSafeHostService,
-  mixpanelTrack,
+  analyticsService,
   userWalletService,
   authenticationService,
   versionService,
@@ -344,12 +346,7 @@ const _recordFetch = async (response, responseData, ...args: Parameters<typeof f
       // Note: functionParams and functionResponse will be added by the calling function
     };
 
-    consoleLog('fetchCallRecorder - response & messageData', response, messageData);
-
-    chrome.runtime.sendMessage({
-      type: 'API_CALL_RECORDED',
-      data: messageData,
-    });
+    consoleInfo('fetchCallRecorder - response & messageData', response, messageData);
   } catch (err) {
     consoleError('Error sending message to UI:', err);
   }
@@ -716,7 +713,7 @@ export class OpenApiService {
   private _loginWithToken = async (userId: string, token: string) => {
     // we shouldn't need to clear storage here anymore
     await authenticationService.signInWithCustomToken(token);
-    await storage.set(CURRENT_ID_KEY, userId);
+    await setLocalData(CURRENT_ID_KEY, userId);
   };
 
   /**
@@ -774,7 +771,7 @@ export class OpenApiService {
 
   register = async (account_key: AccountKeyRequest, username: string) => {
     // Track the time until account_created is called
-    mixpanelTrack.time('account_created');
+    analyticsService.time('account_created');
 
     const config = this.store.config.register;
     const data = await this.sendRequest(
@@ -789,7 +786,7 @@ export class OpenApiService {
     await this._loginWithToken(data.data.id, data.data.custom_token);
 
     // Track the registration
-    mixpanelTrack.track('account_created', {
+    analyticsService.track('account_created', {
       public_key: account_key.public_key,
       sign_algo: getStringFromSignAlgo(account_key.sign_algo),
       hash_algo: getStringFromHashAlgo(account_key.hash_algo),
@@ -1532,7 +1529,7 @@ export class OpenApiService {
     try {
       const response = await fetch(`https://rest-${network}.onflow.org/v1/blocks?height=sealed`);
       const result = await response.json();
-      return result[0].header !== null && result[0].header !== undefined;
+      return Array.isArray(result) && result[0].header !== null && result[0].header !== undefined;
     } catch (err) {
       return false;
     }
@@ -1805,13 +1802,13 @@ export class OpenApiService {
     wallet,
     isChild: ActiveAccountType
   ) => {
-    const loggedInAccounts: LoggedInAccount[] = (await storage.get('loggedInAccounts')) || [];
+    const loggedInAccounts: LoggedInAccount[] = (await getLocalData('loggedInAccounts')) || [];
 
     if (!isChild) {
-      await storage.set('keyIndex', '');
-      await storage.set('hashAlgoString', '');
-      await storage.set('signAlgoString', '');
-      await storage.set('pubKey', '');
+      await setLocalData('keyIndex', '');
+      await setLocalData('hashAlgoString', '');
+      await setLocalData('signAlgoString', '');
+      await setLocalData('pubKey', '');
 
       const { P256, SECP256K1 } = pubKTuple;
 
@@ -1824,10 +1821,10 @@ export class OpenApiService {
           hashAlgoString: keys.keys[0].hashAlgoString,
           publicKey: keys.keys[0].publicKey,
         };
-      await storage.set('keyIndex', keyInfo.index);
-      await storage.set('signAlgoString', keyInfo.signAlgoString);
-      await storage.set('hashAlgoString', keyInfo.hashAlgoString);
-      await storage.set('pubKey', keyInfo.publicKey);
+      await setLocalData('keyIndex', keyInfo.index);
+      await setLocalData('signAlgoString', keyInfo.signAlgoString);
+      await setLocalData('hashAlgoString', keyInfo.hashAlgoString);
+      await setLocalData('pubKey', keyInfo.publicKey);
       // Make sure the address is a FlowAddress
 
       if (!isValidFlowAddress(mainAddress)) {
@@ -1854,7 +1851,7 @@ export class OpenApiService {
       } else {
         loggedInAccounts[accountIndex] = updatedWallet;
       }
-      await storage.set('loggedInAccounts', loggedInAccounts);
+      await setLocalData('loggedInAccounts', loggedInAccounts);
     }
 
     const otherAccounts: LoggedInAccountWithIndex[] = loggedInAccounts
@@ -2041,7 +2038,7 @@ export const getScripts = async (network: string, category: string, scriptName: 
     return modifiedScriptString;
   } catch (error) {
     if (error instanceof Error) {
-      mixpanelTrack.track('script_error', {
+      analyticsService.track('script_error', {
         script_id: scriptName,
         error: error.message,
       });