+++ packages/core/src/service/preference.ts
@@ -6,9 +6,9 @@
  * ******************************************************
  **/
 
+import { getLocalData } from '@onflow/frw-data-model';
 import compareVersions from 'compare-versions';
 
-import storage from '@onflow/frw-extension-shared/storage';
 import { MAINNET_NETWORK, DEFAULT_CURRENCY } from '@onflow/frw-shared/constant';
 import { type FlowNetwork, type Currency } from '@onflow/frw-shared/types';
 
@@ -73,7 +73,7 @@ class PreferenceService {
 
   init = async () => {
     const defaultLang = 'en';
-    const isDeveloperModeEnabled = await storage.get('developerMode');
+    const isDeveloperModeEnabled = await getLocalData<boolean>('developerMode');
     this.store = await createPersistStore<PreferenceStore>({
       name: 'preference',
       template: {
@@ -95,7 +95,7 @@ class PreferenceService {
         firstOpen: false,
         pinnedChain: [],
         // addedToken: {},
-        isDeveloperModeEnabled: isDeveloperModeEnabled || false,
+        isDeveloperModeEnabled: isDeveloperModeEnabled ?? false,
         network: MAINNET_NETWORK,
         isFreeGasFeeEnabled: false,
         displayCurrency: DEFAULT_CURRENCY,
@@ -180,14 +180,6 @@ class PreferenceService {
   setHasOtherProvider = (val: boolean) => {
     this.hasOtherProvider = val;
   };
-
-  // getAcceptLanguages = async () => {
-  //   let langs = await chrome.i18n.getAcceptLanguages();
-  //   if (!langs) langs = [];
-  //   return langs
-  //     .map((lang) => lang.replace(/-/g, '_'))
-  //     .filter((lang) => SUPPORT_LOCALES.includes(lang));
-  // };
 
   getHiddenAddresses = (): PreferenceAccount[] => {
     return structuredClone(this.store.hiddenAddresses);