+++ packages/core/src/service/userWallet.ts
@@ -22,8 +22,9 @@ import {
   registerBatchRefreshListener,
   registerRefreshListener,
   setCachedData,
-  removeUserData,
-  setUserData,
+  getLocalData,
+  removeLocalData,
+  setLocalData,
   activeAccountsKey,
   type ActiveAccountsStore,
   getActiveAccountsData,
@@ -35,8 +36,6 @@ import * as ethUtil from 'ethereumjs-util';
 import { signInAnonymously } from 'firebase/auth/web-extension';
 import { TransactionError } from 'web3';
 
-import { retryOperation } from '@onflow/frw-extension-shared/retryOperation';
-import storage from '@onflow/frw-extension-shared/storage';
 import { DEFAULT_WEIGHT, FLOW_BIP44_PATH } from '@onflow/frw-shared/constant';
 import {
   type PublicPrivateKeyTuple,
@@ -70,8 +69,8 @@ import {
 } from '@onflow/frw-shared/utils';
 
 import { authenticationService } from '.';
+import { analyticsService } from './analytics';
 import keyringService from './keyring';
-import { mixpanelTrack } from './mixpanel';
 import openapiService, { getScripts } from './openapi';
 import preferenceService from './preference';
 import remoteConfigService from './remoteConfig';
@@ -87,6 +86,7 @@ import {
   signWithKey,
 } from '../utils/modules/publicPrivateKey';
 import createPersistStore from '../utils/persistStore';
+import { retryOperation } from '../utils/retryOperation';
 
 interface TransactionNotification {
   url: string;
@@ -275,7 +275,7 @@ class UserWallet {
   // Moved from WalletController to UserWallet
   allowFreeGas = async (): Promise<boolean> => {
     const isFreeGasFeeKillSwitch = await remoteConfigService.getFeatureFlag('free_gas');
-    const isFreeGasFeeEnabled = await storage.get('lilicoPayer');
+    const isFreeGasFeeEnabled = (await getLocalData<boolean>('lilicoPayer')) ?? false;
     return isFreeGasFeeKillSwitch && isFreeGasFeeEnabled;
   };
 
@@ -346,7 +346,7 @@ class UserWallet {
       const mainAccounts = await this.getMainAccounts();
       if (mainAccounts.length === 0) {
         // If the parent address is null, we need to reset to the first parent account
-        await removeUserData(activeAccountsKey(network, pubkey));
+        await removeLocalData(activeAccountsKey(network, pubkey));
         return {
           parentAddress: null,
           currentAddress: null,
@@ -359,7 +359,7 @@ class UserWallet {
         validatedActiveAccounts.currentAddress !== activeAccounts?.currentAddress)
     ) {
       // Only update the active accounts if they have changed and the addresses are not null
-      await setUserData<ActiveAccountsStore>(
+      await setLocalData<ActiveAccountsStore>(
         activeAccountsKey(network, pubkey),
         validatedActiveAccounts
       );
@@ -426,7 +426,7 @@ class UserWallet {
     }
 
     // Save the data in storage
-    await setUserData<ActiveAccountsStore>(activeAccountsKey(network, pubkey), newActiveAccounts);
+    await setLocalData<ActiveAccountsStore>(activeAccountsKey(network, pubkey), newActiveAccounts);
   };
   /**
    * Set the current account - the actively selected account
@@ -732,7 +732,7 @@ class UserWallet {
 
       return txID;
     } catch (error) {
-      mixpanelTrack.track('script_error', {
+      analyticsService.track('script_error', {
         script_id: scriptName,
         error: getErrorMessage(error),
       });
@@ -818,7 +818,7 @@ class UserWallet {
       );
 
       // Track the transaction result
-      mixpanelTrack.track('transaction_result', {
+      analyticsService.track('transaction_result', {
         tx_id: txId,
         is_successful: true,
       });
@@ -907,7 +907,7 @@ class UserWallet {
       });
 
       // Track the transaction error
-      mixpanelTrack.track('transaction_result', {
+      analyticsService.track('transaction_result', {
         tx_id: txId,
         is_successful: false,
         error_message: errorMessage,