+++ packages/core/src/utils/current-id.ts
@@ -1,7 +1,7 @@
-import { CURRENT_ID_KEY, getUserData } from '@onflow/frw-data-model';
+import { CURRENT_ID_KEY, getLocalData } from '@onflow/frw-data-model';
 
 export const returnCurrentProfileId = async (): Promise<string | undefined> => {
-  return await getUserData<string>(CURRENT_ID_KEY);
+  return await getLocalData<string>(CURRENT_ID_KEY);
 };
 
 export const getCurrentProfileId = async (): Promise<string> => {