+++ packages/extension-shared/tsup.config.ts
@@ -3,11 +3,9 @@ import { defineConfig } from 'tsup';
 export default defineConfig({
   entry: {
     'chrome-logger': 'src/chrome-logger.ts',
-    'message/eventBus': 'src/message/eventBus.ts',
     messaging: 'src/messaging.ts',
-    storage: 'src/storage.ts',
-    retryOperation: 'src/retryOperation.ts',
     'contact-utils': 'src/contact-utils.ts',
+    'chrome-storage': 'src/chrome-storage.ts',
   },
   format: ['esm'],
   dts: {