+++ packages/extension-shared/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @onflow/frw-extension-shared
 
+## 0.2.2
+
+### Patch Changes
+
+- 9c75c07: Updated chrome storage export
+
+## 0.2.1
+
+### Patch Changes
+
+- 32eb4bf: Updated exports to work properly on npm
+- Updated dependencies [32eb4bf]
+  - @onflow/frw-shared@0.2.1
+
+## 0.2.0
+
+### Minor Changes
+
+- 1d87bbf: Removed all chrome refrerences and updated data model
+
+### Patch Changes
+
+- Updated dependencies [1d87bbf]
+  - @onflow/frw-shared@0.2.0
+
 ## 0.1.1
 
 ### Patch Changes