+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @onflow/frw-shared
 
+## 0.2.1
+
+### Patch Changes
+
+- 32eb4bf: Updated exports to work properly on npm
+
+## 0.2.0
+
+### Minor Changes
+
+- 1d87bbf: Removed all chrome refrerences and updated data model
+
 ## 0.1.1
 
 ### Patch Changes